% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbcValidationModel.R
\name{AbcValidationModel}
\alias{AbcValidationModel}
\title{R6 Class Representing a Approximate Bayesian Computation (ABC) Model.}
\description{
An R6 class to to represent a model for Approximate Bayesian Computation (ABC) 
validation and best simulation model selection. The class wraps functionality from
the \emph{abc} library, and includes methods for resolving non-finite metrics, 
centering and scaling the ABC inputs, running the ABC analysis, and generating ABC
diagnostics.
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericModel]{paleopop::GenericModel}} -> \code{AbcValidationModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_attributes}}{A vector of public model attribute names - another stored as private.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simulation_parameters}}{A data frame of sample model parameters for each simulation.}

\item{\code{simulation_summary_metrics}}{A data frame of result summary metrics for each simulation.}

\item{\code{observed_metric_targets}}{A vector of observed targets for each summary metric.}

\item{\code{random_seed}}{A seed for randomizing the order of the simulation samples.}

\item{\code{random_indices}}{Randomized simulation sample indices for the ABC inputs and consequently the ABC results.}

\item{\code{non_finite_replacements}}{A list of numeric values or function names (character strings) or direct assignments (assigned or loaded via source paths) for replacing NAs in specified (list names) summary metrics.}

\item{\code{input_center_scale_values}}{A nested list of center and scale values for ABC input parameters/metrics.}

\item{\code{output_dir}}{Directory path for ABC regression diagnostic and other outputs.}

\item{\code{abc_return_object}}{Object returned by the ABC function (see abc documentation).}

\item{\code{selected_simulations}}{A data frame of simulation sample indices and weights selected/assigned by the ABC validation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AbcValidationModel$new()}}
\item \href{#method-save_to_rds}{\code{AbcValidationModel$save_to_rds()}}
\item \href{#method-run_abc}{\code{AbcValidationModel$run_abc()}}
\item \href{#method-resolve_nonfinite_metrics}{\code{AbcValidationModel$resolve_nonfinite_metrics()}}
\item \href{#method-center_scale_abc_inputs}{\code{AbcValidationModel$center_scale_abc_inputs()}}
\item \href{#method-generate_abc_diagnostics}{\code{AbcValidationModel$generate_abc_diagnostics()}}
\item \href{#method-clone}{\code{AbcValidationModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_aliases}{\code{paleopop::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_names">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_names}{\code{paleopop::GenericModel$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attributes">}\href{../../paleopop/html/GenericModel.html#method-get_attributes}{\code{paleopop::GenericModel$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="new_clone">}\href{../../paleopop/html/GenericModel.html#method-new_clone}{\code{paleopop::GenericModel$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="set_attributes">}\href{../../paleopop/html/GenericModel.html#method-set_attributes}{\code{paleopop::GenericModel$set_attributes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets given attributes individually and/or from a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbcValidationModel$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}

\item{\code{template}}{Template population model containing fixed (non-sampled) attributes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_to_rds"></a>}}
\subsection{Method \code{save_to_rds()}}{
Saves a class object to a RDS file via a path to a file, or a directory (using class name).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbcValidationModel$save_to_rds(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to file or directory (where file \emph{<class-name>.RDS} is saved).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_abc"></a>}}
\subsection{Method \code{run_abc()}}{
Pre-processes inputs, runs ABC validation for input parameters, and stores abc (and optionally diagnostic) outputs (see \emph{abc} documentation).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbcValidationModel$run_abc(
  param = NULL,
  sumstat = NULL,
  target = NULL,
  tol = 0.01,
  method = "neuralnet",
  output_diagnostics = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{A data frame of sample model parameters for each simulation.}

\item{\code{sumstat}}{A data frame of result summary metrics for each simulation.}

\item{\code{target}}{A vector of observed targets for each summary metric.}

\item{\code{tol}}{Tolerance or proportion of models to select.}

\item{\code{method}}{ABC algorithm to be applied (default is a neural network algorithm - see \emph{abc} documentation) .}

\item{\code{output_diagnostics}}{Boolean to indicate whether or not to output diagnostics (PDF file - default is FALSE).}

\item{\code{...}}{Additional ABC parameters passed individually (see \emph{abc} documentation).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resolve_nonfinite_metrics"></a>}}
\subsection{Method \code{resolve_nonfinite_metrics()}}{
Attempts to resolve any non-finite simulation summary metric values (and optionally changing them to NAs) via the non finite replacements parameter (a list of values/functions for replacing non-finite values).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbcValidationModel$resolve_nonfinite_metrics(use_nas = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_nas}}{Boolean to indicate whether or not to replace all non-finite values with NAs (default is TRUE).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-center_scale_abc_inputs"></a>}}
\subsection{Method \code{center_scale_abc_inputs()}}{
Centers and scales the model parameters, result summary metrics and observed targets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbcValidationModel$center_scale_abc_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_abc_diagnostics"></a>}}
\subsection{Method \code{generate_abc_diagnostics()}}{
Generates the ABC diagnostics (see \emph{abc} documentation) as a PDF file in the output directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbcValidationModel$generate_abc_diagnostics(output_dir = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_dir}}{Output directory path for the ABC diagnostics PDF file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbcValidationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
