% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrierSeaIceModel.R
\name{BarrierSeaIceModel}
\alias{BarrierSeaIceModel}
\title{R6 Class Representing a Barrier Sea Ice Model.}
\description{
R6 class functionality for modelling sea, ice and other barriers to dispersal within
a spatially-explicit population model. The barrier model calculates distance
multipliers to modify distance-based dispersal rates for simulated migrations in
a spatio-temporal frictional landscape. The frictional landscape is defined via
conductance values, the inverse of friction, which ranges from zero (barrier) to one
(no friction) with values in-between representing some friction. For example, a 
conductance value of 1/5 = 0.2 represents a landscape in which simulated animals move
5 times slower than a non-friction landscape. In this example the resultant distance
multiplier would be 5, thus reducing the effective dispersal range.
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericModel]{paleopop::GenericModel}} -> \code{BarrierSeaIceModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_attributes}}{A list of public model attributes (name-value pairs) - another stored as private.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parallel_cores}}{Number of cores for running the simulations in parallel.}

\item{\code{transition_directions}}{Number of transition directions or neighbours in which cells are connected: usually 4, 8 (default), or 16 (see gdistance package).}

\item{\code{coordinates}}{Data frame (or matrix) of X-Y population coordinates (WGS84) in longitude (degrees West) and latitude (degrees North).}

\item{\code{barrier_sea_ice_matrix}}{Matrix of conductance (inverse friction) values (range: 0 = barrier; 0 < some friction < 1; 1 = no friction) for each grid cell (rows) at each simulation time step (columns).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-calculate_distance_multipliers}{\code{BarrierSeaIceModel$calculate_distance_multipliers()}}
\item \href{#method-clone}{\code{BarrierSeaIceModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="save_to_rds">}\href{../../paleopop/html/GenericClass.html#method-save_to_rds}{\code{paleopop::GenericClass$save_to_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_aliases}{\code{paleopop::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_names">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_names}{\code{paleopop::GenericModel$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attributes">}\href{../../paleopop/html/GenericModel.html#method-get_attributes}{\code{paleopop::GenericModel$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="initialize">}\href{../../paleopop/html/GenericModel.html#method-initialize}{\code{paleopop::GenericModel$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="new_clone">}\href{../../paleopop/html/GenericModel.html#method-new_clone}{\code{paleopop::GenericModel$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="set_attributes">}\href{../../paleopop/html/GenericModel.html#method-set_attributes}{\code{paleopop::GenericModel$set_attributes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_distance_multipliers"></a>}}
\subsection{Method \code{calculate_distance_multipliers()}}{
Calculates and returns spatio-temporal dispersal distance multipliers for each in-range migration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BarrierSeaIceModel$calculate_distance_multipliers(
  dispersal_indices = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dispersal_indices}}{Two-column matrix representing the target and source coordinate index for each in-range migration.}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Temporal list of dispersal distance multiplier arrays with values for each in-range migration.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BarrierSeaIceModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
