% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bison_hs_raster}
\alias{bison_hs_raster}
\title{Bison vignette habitat suitability raster}
\format{
A \emph{raster::RasterStack} object:
\describe{
  \item{dimensions}{21 rows by 180 columns by 1001 layers}
  \item{resolution}{2 by 2 degree grid cells}
  \item{extent}{longitude -180 to 180 degrees; latitude 42 to 84 degrees}
  \item{values}{Estimated habitat suitability values of 0 to 1}
}
}
\source{
TBA
}
\usage{
bison_hs_raster
}
\description{
A \emph{raster} dataset defining estimated habitat suitability values for each grid
cells of the Siberian study region of the bison example vignette.
}
\keyword{datasets}
