% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationModel.R
\name{CorrelationModel}
\alias{CorrelationModel}
\title{R6 Class Representing a Correlation Model.}
\description{
R6 class functionality for modelling correlations within a spatially-explicit
population model. The model calculates correlations between population model
cells using a distance-based function: \emph{a*exp(-distance/b)}, where
\emph{a} (amplitude) and \emph{b} (breadth) are configurable model
attributes. The model also calculates a Cholesky decomposition of the
correlation matrix, which is utilized to generate (optionally temporal)
correlated normal deviates.
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericModel]{paleopop::GenericModel}} -> \code{CorrelationModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_attributes}}{A list of public model attributes (name-value pairs) - another stored as private.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{coordinates}}{Data frame (or matrix) of X-Y population coordinates (WGS84) in longitude (degrees West) and latitude (degrees North).}

\item{\code{correlation_amplitude}}{Correlation function: \emph{a*exp(-distance/b)} \emph{a} parameter. Represents the amplitude or maximum magnitude of correlation values between model cells.}

\item{\code{correlation_breadth}}{Correlation function: \emph{a*exp(-distance/b)} \emph{b} parameter. Represents the breadth of the correlation between model cells. Typically estimated via average distance between correlated population cells.}

\item{\code{correlation_matrix}}{Correlation matrix calculated via correlation function: \emph{a*exp(-distance/b)}.}

\item{\code{t_decomposition_matrix}}{The transposed Cholesky decomposition of the correlation matrix.}

\item{\code{compact_only}}{Boolean to indicate that only the compact versions of matrices will be maintained once calculated.}

\item{\code{t_decomposition_compact_matrix}}{A compact (rows) version of the transposed Cholesky decomposition of the correlation matrix.}

\item{\code{t_decomposition_compact_map}}{A map of the original row populations for the compact transposed decomposition matrix.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CorrelationModel$new()}}
\item \href{#method-calculate_distance_matrix}{\code{CorrelationModel$calculate_distance_matrix()}}
\item \href{#method-calculate_correlations}{\code{CorrelationModel$calculate_correlations()}}
\item \href{#method-calculate_cholesky_decomposition}{\code{CorrelationModel$calculate_cholesky_decomposition()}}
\item \href{#method-calculate_compact_decomposition}{\code{CorrelationModel$calculate_compact_decomposition()}}
\item \href{#method-get_compact_decomposition}{\code{CorrelationModel$get_compact_decomposition()}}
\item \href{#method-generate_correlated_normal_deviates}{\code{CorrelationModel$generate_correlated_normal_deviates()}}
\item \href{#method-clone}{\code{CorrelationModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="save_to_rds">}\href{../../paleopop/html/GenericClass.html#method-save_to_rds}{\code{paleopop::GenericClass$save_to_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_aliases}{\code{paleopop::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_names">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_names}{\code{paleopop::GenericModel$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attributes">}\href{../../paleopop/html/GenericModel.html#method-get_attributes}{\code{paleopop::GenericModel$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="new_clone">}\href{../../paleopop/html/GenericModel.html#method-new_clone}{\code{paleopop::GenericModel$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="set_attributes">}\href{../../paleopop/html/GenericModel.html#method-set_attributes}{\code{paleopop::GenericModel$set_attributes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets given attributes individually and/or from a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$new(compact_only = TRUE, attribute_aliases = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{compact_only}}{Boolean to indicate that only the compact versions of matrices will be maintained once calculated.}

\item{\code{attribute_aliases}}{Optional list of extra alias names for model attributes (form: alias = "attribute") to be used with the set and get attributes methods.}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_distance_matrix"></a>}}
\subsection{Method \code{calculate_distance_matrix()}}{
Returns a matrix with the calculated distance between each pair of population cells (via \emph{coordinates}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$calculate_distance_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Matrix with distances between population cells (\emph{population} rows by \emph{population} columns).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_correlations"></a>}}
\subsection{Method \code{calculate_correlations()}}{
Calculates the correlation matrix by applying the distance-based correlation function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$calculate_correlations(
  distance_matrix = NULL,
  decimals = NULL,
  threshold = 1e-07,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{Optional pre-calculated matrix with distances between population cells (\emph{population} rows by \emph{population} columns).}

\item{\code{decimals}}{Optional number of decimal places for correlation values.}

\item{\code{threshold}}{Optional threshold (minimum value) for correlation values (default 0.0000001).}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_cholesky_decomposition"></a>}}
\subsection{Method \code{calculate_cholesky_decomposition()}}{
Calculates the transposed Cholesky decomposition of the correlation matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$calculate_cholesky_decomposition(
  distance_matrix = NULL,
  decimals = NULL,
  threshold = 1e-07,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{Optional pre-calculated matrix with distances between population cells (population rows by population columns).}

\item{\code{decimals}}{Optional number of decimal places for correlation values.}

\item{\code{threshold}}{Optional threshold (minimum value) for correlation values (default 0.0000001).}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_compact_decomposition"></a>}}
\subsection{Method \code{calculate_compact_decomposition()}}{
Compacts the transposed Cholesky decomposition of the correlation matrix into the minimal number of rows, which are mapped to the original matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$calculate_compact_decomposition(distance_matrix = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{Optional pre-calculated matrix with distances between population cells (population rows by population columns).}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_compact_decomposition"></a>}}
\subsection{Method \code{get_compact_decomposition()}}{
Returns a compact transposed Cholesky decomposition of the correlation and a corresponding map of population indices in a list with names: matrix, map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$get_compact_decomposition(distance_matrix = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{Optional pre-calculated matrix with distances between population cells (population rows by population columns).}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List containing a compact Cholesky decomposition matrix and a corresponding map of population indices (for the compacted rows).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_correlated_normal_deviates"></a>}}
\subsection{Method \code{generate_correlated_normal_deviates()}}{
Generates correlated normal deviates from the correlation model, utilizing the optional random seed and optional temporal correlation across time steps.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$generate_correlated_normal_deviates(
  random_seed = NULL,
  temporal_correlation = 1,
  time_steps = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{random_seed}}{Optional seed for the random generation of correlated deviates.}

\item{\code{temporal_correlation}}{Optional temporal correlation coefficient (0-1; default = 1).}

\item{\code{time_steps}}{Optional number of time steps for temporal correlation (default = 1 or none).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Array (non-temporal) or matrix (temporal) of correlated normal deviates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
