% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{DispersalModel}
\alias{DispersalModel}
\title{R6 Class Representing a Dispersal Model.}
\description{
R6 class functionality for modelling dispersals within a spatially-explicit
population model. The model calculates dispersal rates between population model
cells using a distance-based function: \emph{p*exp(-distance/b)} for
\emph{distance <= r} (otherwise zero), where \emph{p} (proportion), \emph{b}
(breadth or average distance) and \emph{r} (range or maximum distance) are
configurable model attributes. The dispersal rates are adjusted to limit
emigration from each cell to \emph{p}. The model also generates data for
constructing compacted dispersal matrices. It dynamically generates attributes
defined as \emph{outputs} (default: \emph{dispersal_data}) given sampled
\emph{inputs} (default: \emph{dispersal_proportion} and
\emph{dispersal_max_distance}).
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericModel]{paleopop::GenericModel}} -> \code{\link[paleopop:GenerativeModel]{paleopop::GenerativeModel}} -> \code{DispersalModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_attributes}}{A vector of public model attribute names - another stored as private.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{coordinates}}{Data frame (or matrix) of X-Y population coordinates (WGS84) in longitude (degrees West) and latitude (degrees North).}

\item{\code{distance_classes}}{Vector of distance interval boundaries (in km) for calculating discrete dispersal rates.}

\item{\code{distance_data}}{Data frame of distance classes including indices for the construction of compact matrices (columns: target_pop, source_pop, compact_row, distance_class).}

\item{\code{dispersal_function_data}}{Data frame of discrete dispersal function values. Optional first column may provide distance intervals (non-inclusive lower bounds).}

\item{\code{decimals}}{Optional number of decimal places applied to the calculated dispersal rates (default: NULL = no rounding).}

\item{\code{proportion_multiplier}}{Optional constant multiplier to be applied to the dispersal function \emph{p} parameter. May be used to represent the fraction of the population capable of migrating.}

\item{\code{dispersal_proportion}}{Dispersal function: \emph{p*exp(-distance/b)} \emph{p} parameter. Represents the proportion and limit of dispersers between model cells.}

\item{\code{dispersal_breadth}}{Dispersal function: \emph{p*exp(-distance/b)} \emph{b} parameter. Represents the breadth of the dispersal between model cells. Typically estimated via average migration distance.}

\item{\code{dispersal_max_distance}}{Dispersal maximum distance or range (\emph{r}) parameter limits the use of the dispersal function: \emph{p*exp(-distance/b)}. The function is utilized when \emph{distance <= r} otherwise the dispersal rate is set to zero.}

\item{\code{dispersal_index}}{Sampled index for the dispersal function data frame (to look-up dispersal function parameters).}

\item{\code{dispersal_matrix}}{Dispersal matrix calculated via dispersal function.}

\item{\code{dispersal_data}}{Data frame of non-zero dispersal rates including indices for the construction of compact matrices (columns: target_pop, source_pop, emigrant_row, immigrant_row, dispersal_rate).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DispersalModel$new()}}
\item \href{#method-generative_requirements_satisfied}{\code{DispersalModel$generative_requirements_satisfied()}}
\item \href{#method-set_distance_classes}{\code{DispersalModel$set_distance_classes()}}
\item \href{#method-calculate_distance_matrix}{\code{DispersalModel$calculate_distance_matrix()}}
\item \href{#method-calculate_distance_data}{\code{DispersalModel$calculate_distance_data()}}
\item \href{#method-calculate_dispersals}{\code{DispersalModel$calculate_dispersals()}}
\item \href{#method-clone}{\code{DispersalModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="save_to_rds">}\href{../../paleopop/html/GenericClass.html#method-save_to_rds}{\code{paleopop::GenericClass$save_to_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_aliases}{\code{paleopop::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_names">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_names}{\code{paleopop::GenericModel$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attributes">}\href{../../paleopop/html/GenericModel.html#method-get_attributes}{\code{paleopop::GenericModel$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="set_attributes">}\href{../../paleopop/html/GenericModel.html#method-set_attributes}{\code{paleopop::GenericModel$set_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="add_file_template">}\href{../../paleopop/html/GenerativeModel.html#method-add_file_template}{\code{paleopop::GenerativeModel$add_file_template()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="add_function_template">}\href{../../paleopop/html/GenerativeModel.html#method-add_function_template}{\code{paleopop::GenerativeModel$add_function_template()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="add_generative_requirements">}\href{../../paleopop/html/GenerativeModel.html#method-add_generative_requirements}{\code{paleopop::GenerativeModel$add_generative_requirements()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="generate">}\href{../../paleopop/html/GenerativeModel.html#method-generate}{\code{paleopop::GenerativeModel$generate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="new_clone">}\href{../../paleopop/html/GenerativeModel.html#method-new_clone}{\code{paleopop::GenerativeModel$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="read_file">}\href{../../paleopop/html/GenerativeModel.html#method-read_file}{\code{paleopop::GenerativeModel$read_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="run_function">}\href{../../paleopop/html/GenerativeModel.html#method-run_function}{\code{paleopop::GenerativeModel$run_function()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets the generative template and requirements as well as any attributes passed via a \emph{params} list or individually.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalModel$new(
  generative_template = NULL,
  generative_requirements = NULL,
  attribute_aliases = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generative_template}}{Optional nested object for generative dispersal model attributes that need to be maintained when a new clone object is generated for a sample simulation.}

\item{\code{generative_requirements}}{Optional list of attribute names and the template setting (\emph{"file"} or \emph{"function"}) that is required to generate their values (otherwise default functionality is used).}

\item{\code{attribute_aliases}}{Optional list of extra alias names for model attributes (form: alias = "attribute") to be used with the set and get attributes methods.}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generative_requirements_satisfied"></a>}}
\subsection{Method \code{generative_requirements_satisfied()}}{
Returns a boolean to indicate that all the default, file and/or function template settings that are required for attribute generation are present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalModel$generative_requirements_satisfied()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean to indicate that the required settings for attribute generation are present.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_distance_classes"></a>}}
\subsection{Method \code{set_distance_classes()}}{
Sets the distance classes to a sequence of values from minimum to maximum in steps of interval size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalModel$set_distance_classes(
  minimum = 1,
  maximum = 10,
  interval = (maximum - minimum)/10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minimum}}{Minimum or first distance class sequence value (default = 1).}

\item{\code{maximum}}{Maximum or last distance class value (default = 10).}

\item{\code{interval}}{Interval or distance class sequence step size  (default = (maximum - minimum)/10).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_distance_matrix"></a>}}
\subsection{Method \code{calculate_distance_matrix()}}{
Returns a matrix with the calculated distance between each pair of population cells (via \emph{coordinates}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalModel$calculate_distance_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Matrix with distances between population cells (\emph{population} rows by \emph{population} columns).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_distance_data"></a>}}
\subsection{Method \code{calculate_distance_data()}}{
Calculates the distance class for within-range populations using the set/provided distance classes. Also calculates indices for constructing compact matrices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalModel$calculate_distance_data(distance_matrix = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{Optional pre-calculated matrix with distances between population cells (population rows by population columns).}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_dispersals"></a>}}
\subsection{Method \code{calculate_dispersals()}}{
Calculates a dispersal matrix, or data frame of dispersal rates and compact indices (default), using the conditional dispersal function with dispersal limiting for a simulation sample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalModel$calculate_dispersals(type = "data")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Optional type selector (\emph{"data"} or \emph{"matrix"}) to determine whether to calculate a dispersal matrix or data frame (default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns character string message when calculation prerequisites are not met (for simulation logging).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
