% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalTemplate.R
\name{DispersalTemplate}
\alias{DispersalTemplate}
\title{R6 Class Representing a Nested Container for Dispersal Model Attributes}
\description{
R6 class representing a nested container for generative model attributes that are
maintained when new model clones are created. The container maintains input and
output attribute names, file and function templates, and dispersal model attributes
that need to be maintained when a new clone object is generated for a sample
simulation.
}
\section{Super class}{
\code{\link[paleopop:GenerativeTemplate]{paleopop::GenerativeTemplate}} -> \code{DispersalTemplate}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{barrier_sea_ice_model}}{A BarrierSeaIceModel object for dispersal distance multiplier data.}

\item{\code{coordinates}}{Data frame (or matrix) of X-Y population coordinates (WGS84) in longitude (degrees West) and latitude (degrees North).}

\item{\code{distance_classes}}{Vector of distance interval boundaries (in km) for calculating discrete dispersal rates.}

\item{\code{distance_data}}{Data frame of distance classes including indices for the construction of compact matrices (columns: target_pop, source_pop, compact_row, distance_class).}

\item{\code{dispersal_function_data}}{Data frame of discrete dispersal function values. Optional first column may provide distance intervals (non-inclusive lower bounds).}

\item{\code{decimals}}{Optional number of decimal places applied to the calculated dispersal rates (default: NULL = no rounding).}

\item{\code{proportion_multiplier}}{Optional constant multiplier to be applied to the dispersal function \emph{p} parameter. May be used to represent the fraction of the population capable of migrating.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{DispersalTemplate$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeTemplate" data-id="initialize">}\href{../../paleopop/html/GenerativeTemplate.html#method-initialize}{\code{paleopop::GenerativeTemplate$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DispersalTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
