% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerativeTemplate.R
\name{GenerativeTemplate}
\alias{GenerativeTemplate}
\title{R6 Class Representing a Nested Container for Generative Model Attributes}
\description{
R6 class representing a nested container for generative model attributes that are
maintained when new model clones are created. The container maintains input and
output attribute names, file and function templates, and any inherited class model
attributes that need to be maintained when cloning.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{description}}{A brief description of what the generative model generates.}

\item{\code{inputs}}{An array of input attribute names for the generative model.}

\item{\code{outputs}}{An array of output attribute names for the generative model.}

\item{\code{file_templates}}{A list of file template attributes.}

\item{\code{function_templates}}{A list of function template attributes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenerativeTemplate$new()}}
\item \href{#method-clone}{\code{GenerativeTemplate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method initializes the generative templates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenerativeTemplate$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenerativeTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
