% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericClass.R
\name{GenericClass}
\alias{GenericClass}
\title{R6 Class with Generic Reusable Functionality}
\description{
R6 class with generic (abstract) new cloning and RDS save and read functionality.
}
\examples{
TestInheritedClass <- R6::R6Class("TestInheritedClass",
                                  inherit = GenericClass,
                                  public = list(attr1 = NULL, attr2 = NULL))
test_class <- TestInheritedClass$new()
test_class$attr1 <- 11
test_class$attr2 <- 22
test_class
test_class$new_clone()

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new_clone}{\code{GenericClass$new_clone()}}
\item \href{#method-read_from_rds}{\code{GenericClass$read_from_rds()}}
\item \href{#method-save_to_rds}{\code{GenericClass$save_to_rds()}}
\item \href{#method-clone}{\code{GenericClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new_clone"></a>}}
\subsection{Method \code{new_clone()}}{
Creates a new (re-initialized) object of the current (inherited) object class with optionally passed parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericClass$new_clone(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed via the inherited class constructor (defined in initialize and run via new).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
New object of the inherited class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_from_rds"></a>}}
\subsection{Method \code{read_from_rds()}}{
Reads a class object from a RDS file via a path to a file, or a directory (using class name).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericClass$read_from_rds(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to file or directory containing \emph{<class-name>.RDS} file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_to_rds"></a>}}
\subsection{Method \code{save_to_rds()}}{
Saves a class object to a RDS file via a path to a file, or a directory (using class name).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericClass$save_to_rds(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to file or directory (where file \emph{<class-name>.RDS} is saved).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
