% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericResultsManager.R
\name{GenericResultsManager}
\alias{GenericResultsManager}
\title{R6 Class Representing a Generic Results Manager.}
\description{
An R6 class to represent a generic (abstract) manager for generating or processing of
simulation results, as well as optionally re-generating niche carrying capacities
and/or human densities.
}
\section{Super class}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{GenericResultsManager}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attached}}{A list of dynamically attached attributes (name-value pairs).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sample_data}}{A data frame of sampled parameters for each simulation.}

\item{\code{niche_k_model}}{A niche carrying capacity generative model (\emph{NicheCarryingCapacityModel} or inherited class) object for generating sample carrying capacities.}

\item{\code{human_density_model}}{A human density generative model (\emph{HumanDensityModel} or inherited class) object for generating sample human densities.}

\item{\code{parallel_cores}}{Number of cores for running the simulations in parallel.}

\item{\code{results_dir}}{Results directory path.}

\item{\code{results_filename_attributes}}{A vector of: prefix (optional); attribute names (from the sample data frame); postfix (optional); utilized to construct results filenames.}

\item{\code{error_messages}}{A vector of error messages encountered.}

\item{\code{warning_messages}}{A vector of warning messages encountered.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GenericResultsManager$new()}}
\item \href{#method-save_to_rds}{\code{GenericResultsManager$save_to_rds()}}
\item \href{#method-get_attribute}{\code{GenericResultsManager$get_attribute()}}
\item \href{#method-get_message_sample}{\code{GenericResultsManager$get_message_sample()}}
\item \href{#method-get_results_filename}{\code{GenericResultsManager$get_results_filename()}}
\item \href{#method-clone}{\code{GenericResultsManager$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="new_clone">}\href{../../paleopop/html/GenericClass.html#method-new_clone}{\code{paleopop::GenericClass$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets any included attributes (sample_data, niche_k_model, human_density_model, parallel_cores, results_dir, results_filename_attributes) and attaches other attributes individually listed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericResultsManager$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters listed individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_to_rds"></a>}}
\subsection{Method \code{save_to_rds()}}{
Saves a model to a RDS file via a path to a file, or a directory (using class name), or to a results directory (if set) when path is not provided.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericResultsManager$save_to_rds(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to file or directory (where file \emph{<class-name>.RDS} is saved).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attribute"></a>}}
\subsection{Method \code{get_attribute()}}{
Returns a named manager or attached attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericResultsManager$get_attribute(param)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Character string name of the attribute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Selected attribute value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_message_sample"></a>}}
\subsection{Method \code{get_message_sample()}}{
Substitutes the specified sample details into a status message (using sprintf) and returns the result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericResultsManager$get_message_sample(status_message, sample_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status_message}}{Character string message with a \emph{\%s} placeholder for sample details.}

\item{\code{sample_index}}{Row index of sample data frame containing details of substitution parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Status message with substituted sample details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_results_filename"></a>}}
\subsection{Method \code{get_results_filename()}}{
Constructs and returns the results filename based on the sample data frame index and results filename attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericResultsManager$get_results_filename(sample_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_index}}{Row index of sample data frame containing details of substitution parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Results filename with substituted sample details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericResultsManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
