% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDensityModel.R
\name{HumanDensityModel}
\alias{HumanDensityModel}
\title{R6 Class Representing a Human Density Model.}
\description{
R6 class functionality for modelling human densities within a spatially-explicit
population model. The model generates samples of human densities as a fraction of
maximum capacity via a triangular distribution based on configured human density
means and standard deviations for each simulation cell, as well as optional limits.
It dynamically generates attributes defined as \emph{outputs} (default:
\emph{human_densities}) given sampled \emph{inputs} (default:
\emph{human_density_sample}).
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericModel]{paleopop::GenericModel}} -> \code{\link[paleopop:GenerativeModel]{paleopop::GenerativeModel}} -> \code{HumanDensityModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_attributes}}{A list of public model attributes (name-value pairs) - another stored as private.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{carrying_capacity_mean}}{Matrix (or data frame) of human carrying capacity time-series data means (simulation cells by duration).}

\item{\code{carrying_capacity_sd}}{Matrix (or data frame) of human carrying capacity time-series data standard deviations (simulation cells by duration).}

\item{\code{human_occupancy_mask}}{Optional binary mask (matrix or data frame) for human carrying capacity time-series data (simulation cells by duration).}

\item{\code{lower_threshold}}{Optional lower threshold for human occupancy (lower values are set to zero).}

\item{\code{mean_upper}}{Optional upper limit for sample distribution mean carrying capacities.}

\item{\code{max_upper}}{Optional upper limit for sample distribution maximum carrying capacities.}

\item{\code{sd_number}}{Number of standard deviations from the sample distribution mean to the lower and upper limits (default = 1).}

\item{\code{distrib_data}}{Calculated sample distribution minimum, maximum and mean carrying capacity time-series data.}

\item{\code{uses_correlations}}{A boolean to indicate that a correlation model is used for generating correlated random deviates.}

\item{\code{correlation_model}}{A correlation model for generating correlated random deviates (type of \emph{CorrelationModel} or an inherited class).}

\item{\code{temporal_correlation}}{Absolute correlation coefficient between simulation time steps for all grid cells (0-1; default = 1).}

\item{\code{decimals}}{Number of decimal places applied to the calculated human densities (default: NULL = no rounding).}

\item{\code{distrib_mean_matrix}}{Calculated sample distribution mean carrying capacity time-series matrix.}

\item{\code{human_density_sample}}{Human density sample number (or file index) or probability (for default generation).}

\item{\code{human_density_lower}}{Human density lower probability for correlated generation.}

\item{\code{human_density_upper}}{Human density upper probability for correlated generation.}

\item{\code{human_density_seed}}{Human density random seed for repeatable correlated generation.}

\item{\code{human_densities}}{Calculated sample human density values (simulation cells by duration).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HumanDensityModel$new()}}
\item \href{#method-generative_requirements_satisfied}{\code{HumanDensityModel$generative_requirements_satisfied()}}
\item \href{#method-calculate_sampling_parameters}{\code{HumanDensityModel$calculate_sampling_parameters()}}
\item \href{#method-calculate_human_densities}{\code{HumanDensityModel$calculate_human_densities()}}
\item \href{#method-clone}{\code{HumanDensityModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="save_to_rds">}\href{../../paleopop/html/GenericClass.html#method-save_to_rds}{\code{paleopop::GenericClass$save_to_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_aliases}{\code{paleopop::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_names">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_names}{\code{paleopop::GenericModel$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attributes">}\href{../../paleopop/html/GenericModel.html#method-get_attributes}{\code{paleopop::GenericModel$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="set_attributes">}\href{../../paleopop/html/GenericModel.html#method-set_attributes}{\code{paleopop::GenericModel$set_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="add_file_template">}\href{../../paleopop/html/GenerativeModel.html#method-add_file_template}{\code{paleopop::GenerativeModel$add_file_template()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="add_function_template">}\href{../../paleopop/html/GenerativeModel.html#method-add_function_template}{\code{paleopop::GenerativeModel$add_function_template()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="add_generative_requirements">}\href{../../paleopop/html/GenerativeModel.html#method-add_generative_requirements}{\code{paleopop::GenerativeModel$add_generative_requirements()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="generate">}\href{../../paleopop/html/GenerativeModel.html#method-generate}{\code{paleopop::GenerativeModel$generate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="new_clone">}\href{../../paleopop/html/GenerativeModel.html#method-new_clone}{\code{paleopop::GenerativeModel$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="read_file">}\href{../../paleopop/html/GenerativeModel.html#method-read_file}{\code{paleopop::GenerativeModel$read_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeModel" data-id="run_function">}\href{../../paleopop/html/GenerativeModel.html#method-run_function}{\code{paleopop::GenerativeModel$run_function()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets the generative template and requirements as well as any attributes passed via a \emph{params} list or individually.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumanDensityModel$new(
  generative_template = NULL,
  generative_requirements = NULL,
  attribute_aliases = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generative_template}}{Optional nested object for generative human density model attributes that need to be maintained when a new clone object is generated for a sample simulation.}

\item{\code{generative_requirements}}{Optional list of attribute names and the template setting (\emph{"file"} or \emph{"function"}) that is required to generate their values (otherwise default functionality is used).}

\item{\code{attribute_aliases}}{Optional list of extra alias names for model attributes (form: alias = "attribute") to be used with the set and get attributes methods.}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generative_requirements_satisfied"></a>}}
\subsection{Method \code{generative_requirements_satisfied()}}{
Returns a boolean to indicate that all the default, file and/or function template settings that are required for attribute generation are present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumanDensityModel$generative_requirements_satisfied()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean to indicate that the required settings for attribute generation are present.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_sampling_parameters"></a>}}
\subsection{Method \code{calculate_sampling_parameters()}}{
Calculates the minimum, mean and maximum human density values for each simulation grid and time step based on the input mean and standard deviation data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumanDensityModel$calculate_sampling_parameters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_human_densities"></a>}}
\subsection{Method \code{calculate_human_densities()}}{
Calculates the human density values via the (triangular) sampling distribution for each simulation grid and time step utilizing correlated deviates for each grid cell.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumanDensityModel$calculate_human_densities()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns character string message when calculation prerequisites are not met (for simulation logging).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumanDensityModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
