% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDensityTemplate.R
\name{HumanDensityTemplate}
\alias{HumanDensityTemplate}
\title{R6 Class Representing a Nested Container for Human Density Model Attributes}
\description{
R6 class representing a nested container for human density model attributes that are
maintained when new model clones are created. The container maintains input and
output attribute names, file and function templates, and human density model
attributes that need to be maintained when a new clone object is generated for a
sample simulation.
}
\section{Super class}{
\code{\link[paleopop:GenerativeTemplate]{paleopop::GenerativeTemplate}} -> \code{HumanDensityTemplate}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{carrying_capacity_mean}}{Matrix (or data frame) of human carrying capacity time-series data means (simulation cells by duration).}

\item{\code{carrying_capacity_sd}}{Matrix (or data frame) of human carrying capacity time-series data standard deviations (simulation cells by duration).}

\item{\code{human_occupancy_mask}}{Optional binary mask (matrix or data frame) for human carrying capacity time-series data (simulation cells by duration).}

\item{\code{lower_threshold}}{Optional lower threshold for human occupancy (lower values are set to zero).}

\item{\code{mean_upper}}{Optional upper limit for sample distribution mean carrying capacities.}

\item{\code{max_upper}}{Optional upper limit for sample distribution maximum carrying capacities.}

\item{\code{sd_number}}{Number of standard deviations from the sample distribution mean to the lower and upper limits (default = 1).}

\item{\code{distrib_data}}{Calculated sample distribution minimum, maximum and mean carrying capacity time-series data, combined, scaled, and zero sets eliminated for generation efficiency.}

\item{\code{uses_correlations}}{A boolean to indicate that a correlation model is used for generating correlated random deviates.}

\item{\code{correlation_model}}{A correlation model for generating correlated random deviates (type of \emph{CorrelationModel} or an inherited class).}

\item{\code{temporal_correlation}}{Absolute correlation coefficient between simulation time steps for all grid cells (0-1; default = 1).}

\item{\code{decimals}}{Number of decimal places applied to the calculated human densities (default: NULL = no rounding).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HumanDensityTemplate$new()}}
\item \href{#method-clone}{\code{HumanDensityTemplate$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets default values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumanDensityTemplate$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HumanDensityTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
