% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSimulationManager.R
\name{MultiSimulationManager}
\alias{MultiSimulationManager}
\title{R6 Class Representing a Multi-Simulation Manager.}
\description{
An R6 class to represent a manager for running multiple population model simulations
and saving results.
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericResultsManager]{paleopop::GenericResultsManager}} -> \code{MultiSimulationManager}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model_template}}{A PopulationModel (or inherited class) object with parameters common to all simulations.}

\item{\code{nested_model}}{A PopulationModel (or inherited class) object with empty sample parameters and a nested model template common to all simulations.}

\item{\code{dispersal_model}}{A dispersal generative model (\emph{DispersalModel} or inherited class) object for generating sample dispersals.}

\item{\code{population_simulator}}{A PopulationSimulator (or inherited class) object for running the simulations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiSimulationManager$new()}}
\item \href{#method-run}{\code{MultiSimulationManager$run()}}
\item \href{#method-set_model_sample}{\code{MultiSimulationManager$set_model_sample()}}
\item \href{#method-log_simulation}{\code{MultiSimulationManager$log_simulation()}}
\item \href{#method-clone}{\code{MultiSimulationManager$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="new_clone">}\href{../../paleopop/html/GenericClass.html#method-new_clone}{\code{paleopop::GenericClass$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="get_attribute">}\href{../../paleopop/html/GenericResultsManager.html#method-get_attribute}{\code{paleopop::GenericResultsManager$get_attribute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="get_message_sample">}\href{../../paleopop/html/GenericResultsManager.html#method-get_message_sample}{\code{paleopop::GenericResultsManager$get_message_sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="get_results_filename">}\href{../../paleopop/html/GenericResultsManager.html#method-get_results_filename}{\code{paleopop::GenericResultsManager$get_results_filename()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="save_to_rds">}\href{../../paleopop/html/GenericResultsManager.html#method-save_to_rds}{\code{paleopop::GenericResultsManager$save_to_rds()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets any included attributes (sample_data, model_template, niche_k_model, dispersal_model, human_density_model, population_simulator, parallel_cores, results_dir, results_filename_attributes) and attaches other attributes individually listed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSimulationManager$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters listed individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
Runs the multiple population simulations (via the set function), stores the results, and creates/writes a simulation log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSimulationManager$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Simulator log as a list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_model_sample"></a>}}
\subsection{Method \code{set_model_sample()}}{
Sets the model sample attributes via the sample data frame, and the niche K, dispersal and human density generative models.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSimulationManager$set_model_sample(model, sample_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Population model (clone) to receive sample attributes.}

\item{\code{sample_index}}{Index of sample from data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_simulation"></a>}}
\subsection{Method \code{log_simulation()}}{
Summarizes the simulation log generated within the run method and writes it to (text and RDS) files in the results directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSimulationManager$log_simulation(simulation_log)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulation_log}}{Nested list of simulation log entries generated via the run method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiSimulationManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
