% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheCarryingCapacityTemplate.R
\name{NicheCarryingCapacityTemplate}
\alias{NicheCarryingCapacityTemplate}
\title{R6 Class Representing a Nested Container for Niche Carrying Capacity Model Attributes}
\description{
R6 class representing a nested container for generative model attributes that are
maintained when new model clones are created. The container maintains input and
output attribute names, file and function templates, and niche carrying capacity
model attributes that need to be maintained when a new clone object is generated
for a sample simulation.
}
\section{Super class}{
\code{\link[paleopop:GenerativeTemplate]{paleopop::GenerativeTemplate}} -> \code{NicheCarryingCapacityTemplate}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{niche_occupancy_mask}}{Optional binary mask (matrix or data frame) for niche carrying capacity time-series data (simulation cells by duration).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{NicheCarryingCapacityTemplate$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenerativeTemplate" data-id="initialize">}\href{../../paleopop/html/GenerativeTemplate.html#method-initialize}{\code{paleopop::GenerativeTemplate$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NicheCarryingCapacityTemplate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
