% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleopop-package.R
\docType{package}
\name{paleopop-package}
\alias{paleopop-package}
\title{paleopop: Pattern-Oriented Modeling Framework for Coupled Niche-Population Paleo-Climatic Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This extension of the poems pattern-oriented modeling (POM) framework provides a collection of modules and functions customized for paleontological time-scales, and optimized for single-generation transitions and large populations, across multiple generations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlobalEcologyLab/paleopop/}
  \item Report bugs at \url{https://github.com/GlobalEcologyLab/paleopop/issues}
}

}
\author{
\strong{Maintainer}: July Pilowsky \email{pilowskyj@caryinstitute.org} (\href{https://orcid.org/0000-0002-6376-2585}{ORCID})

Authors:
\itemize{
  \item Sean Haythorne \email{sean.haythorne@adelaide.edu.au}
  \item Stuart Brown \email{s.brown@adelaide.edu.au} (\href{https://orcid.org/0000-0002-0669-1418}{ORCID})
  \item Damien Fordham \email{damien.fordham@adelaide.edu.au} (\href{https://orcid.org/0000-0003-2137-5592}{ORCID})
}

}
\keyword{internal}
