% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleoPopModel.R
\name{PaleoPopModel}
\alias{PaleoPopModel}
\title{R6 Class Representing a Population Model for the Default Paleopop Simulator}
\description{
R6 class representing a spatially-explicit demographic-based population model with
attributes utilized by the default paleopop simulator function.
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericModel]{paleopop::GenericModel}} -> \code{\link[paleopop:PopulationModel]{paleopop::PopulationModel}} -> \code{PaleoPopModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_attributes}}{A vector of public model attribute names - another stored as private.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{years_per_step}}{Number of years per time step.}

\item{\code{transition_rate}}{Rate of transition (or fecundity) between generations.}

\item{\code{standard_deviation}}{Standard deviation applied to transition rates.}

\item{\code{growth_rate_max}}{Maximum growth rate.}

\item{\code{local_threshold}}{Abundance threshold (that needs to be exceeded) for each population to persist.}

\item{\code{dispersal_target_k_threshold}}{Target population carrying capacity threshold for density dependent dispersal.}

\item{\code{dispersal_data}}{Data frame of non-zero dispersal rates and indices for the construction of compact matrices (columns: target_pop, source_pop, emigrant_row, immigrant_row, dispersal_rate).}

\item{\code{compact_decomposition}}{List containing a pre-calculated compact transposed (Cholesky) decomposition matrix and a corresponding map of population indices for environmental correlation (list names: matrix, map).}

\item{\code{harvest}}{Boolean for utilizing harvesting.}

\item{\code{harvest_max}}{Proportion harvested per year (annual time scale - not generational).}

\item{\code{harvest_g}}{The "G" parameter in the harvest function.}

\item{\code{harvest_z}}{The "Z" parameter in the harvest function.}

\item{\code{harvest_max_n}}{Maximum density per grid cell.}

\item{\code{results_selection}}{List of results selection from ("abundance", "ema", "extirpation", "harvested", "occupancy").}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PaleoPopModel$new()}}
\item \href{#method-is_consistent}{\code{PaleoPopModel$is_consistent()}}
\item \href{#method-is_complete}{\code{PaleoPopModel$is_complete()}}
\item \href{#method-list_completeness}{\code{PaleoPopModel$list_completeness()}}
\item \href{#method-incomplete_attributes}{\code{PaleoPopModel$incomplete_attributes()}}
\item \href{#method-clone}{\code{PaleoPopModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="save_to_rds">}\href{../../paleopop/html/GenericClass.html#method-save_to_rds}{\code{paleopop::GenericClass$save_to_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_aliases}{\code{paleopop::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attributes">}\href{../../paleopop/html/GenericModel.html#method-get_attributes}{\code{paleopop::GenericModel$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="new_clone">}\href{../../paleopop/html/GenericModel.html#method-new_clone}{\code{paleopop::GenericModel$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="set_attributes">}\href{../../paleopop/html/GenericModel.html#method-set_attributes}{\code{paleopop::GenericModel$set_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="PopulationModel" data-id="get_attribute_names">}\href{../../paleopop/html/PopulationModel.html#method-get_attribute_names}{\code{paleopop::PopulationModel$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="PopulationModel" data-id="set_sample_attributes">}\href{../../paleopop/html/PopulationModel.html#method-set_sample_attributes}{\code{paleopop::PopulationModel$set_sample_attributes()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets template model and sets given attributes individually and/or from a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopModel$new(template = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{template}}{Template population model containing fixed (non-sampled) attributes.}

\item{\code{...}}{Parameters passed via a \emph{params} list or individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_consistent"></a>}}
\subsection{Method \code{is_consistent()}}{
Returns a boolean to indicate if (optionally selected or all) model attributes (such as dimensions) are consistent.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopModel$is_consistent(params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Optional array of parameter/attribute names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean to indicate consistency of selected/all attributes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_complete"></a>}}
\subsection{Method \code{is_complete()}}{
Returns a boolean to indicate if all attributes necessary to simulate the model have been set and are consistent.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopModel$is_complete()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean to indicate model completeness (and consistency).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_completeness"></a>}}
\subsection{Method \code{list_completeness()}}{
Returns a list of booleans for each parameter to indicate which attributes necessary to simulate the model have been set and are consistent.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopModel$list_completeness()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of booleans for each parameter to indicate to indicate completeness (and consistency).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-incomplete_attributes"></a>}}
\subsection{Method \code{incomplete_attributes()}}{
Returns a list of attributes necessary to simulate the model that are incomplete.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopModel$incomplete_attributes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of incomplete attributes which prevent the model simulation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
