% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleoPopResultsModel.R
\name{PaleoPopResultsModel}
\alias{PaleoPopResultsModel}
\title{R6 Class Representing a Results Model for the Default Paleopop Simulator.}
\description{
An R6 class to represent a generic (abstract) model for encapsulating and dynamically
generating \emph{paleopop} simulation results associated with the default paleopop 
simulator function, as well as optional re-generated niche carrying capacities and/or
human densities.
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericModel]{paleopop::GenericModel}} -> \code{\link[paleopop:GenericResultsModel]{paleopop::GenericResultsModel}} -> \code{PaleoPopResultsModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_attributes}}{An array of public model attribute names - another stored as private.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{default}}{Default value(s) utilized when applying primitive metric functions (e.g. max) to the results.}

\item{\code{coordinates}}{Data frame (or matrix) of X-Y population coordinates (WGS84) in longitude (degrees West) and latitude (degrees North).}

\item{\code{abundance}}{Matrix of population abundance across simulation time-steps (population rows by duration columns).}

\item{\code{abundance_trend}}{Trend or average (Sen's) slope of total abundance (optionally across a time-step interval).}

\item{\code{ema}}{Matrix of population expected minimum abundance (EMA) across simulation time-steps (population rows by duration columns).}

\item{\code{extirpation}}{Array of population extirpation times.}

\item{\code{extinction_location}}{The weighted centroid of cells occupied in the time-step prior to the extirpation of all populations (if occurred).}

\item{\code{harvested}}{Matrix of the number of animals harvested from each population at each time-step (population rows by duration columns).}

\item{\code{occupancy}}{Array of the number of populations occupied at each time-step.}

\item{\code{occupancy_mask}}{Optional binary mask matrix for each population at each time-step (population rows by duration columns).}

\item{\code{trend_interval}}{Optional time-step range (indices) for trend calculations (assumes indices begin after the burn-in when utilized).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new_clone}{\code{PaleoPopResultsModel$new_clone()}}
\item \href{#method-clone}{\code{PaleoPopResultsModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="save_to_rds">}\href{../../paleopop/html/GenericClass.html#method-save_to_rds}{\code{paleopop::GenericClass$save_to_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attribute_aliases">}\href{../../paleopop/html/GenericModel.html#method-get_attribute_aliases}{\code{paleopop::GenericModel$get_attribute_aliases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="get_attributes">}\href{../../paleopop/html/GenericModel.html#method-get_attributes}{\code{paleopop::GenericModel$get_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericModel" data-id="set_attributes">}\href{../../paleopop/html/GenericModel.html#method-set_attributes}{\code{paleopop::GenericModel$set_attributes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsModel" data-id="get_attribute_names">}\href{../../paleopop/html/GenericResultsModel.html#method-get_attribute_names}{\code{paleopop::GenericResultsModel$get_attribute_names()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsModel" data-id="initialize">}\href{../../paleopop/html/GenericResultsModel.html#method-initialize}{\code{paleopop::GenericResultsModel$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new_clone"></a>}}
\subsection{Method \code{new_clone()}}{
Creates a new (re-initialized) object of the current (inherited) object class with optionally passed parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopResultsModel$new_clone(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed via the inherited class constructor (defined in initialize and run via new).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
New object of the inherited class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PaleoPopResultsModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
