% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationSimulator.R
\name{PopulationSimulator}
\alias{PopulationSimulator}
\title{R6 Class Representing a Population Simulator.}
\description{
An R6 class for running individual population model simulations via a simulation
function, storing results, and generating success/error statuses.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attached}}{A list of attached attributes (name-value pairs).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{population_model}}{A PopulationModel object or an inherited class object.}

\item{\code{simulation_function}}{Name (character string) or direct assignment (assigned or loaded via source path) of the population simulation function, which takes a population model as an input and returns the simulation results.}

\item{\code{sample_id}}{An identifier for the simulation sample.}

\item{\code{results}}{A list of result structures.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PopulationSimulator$new()}}
\item \href{#method-new_clone}{\code{PopulationSimulator$new_clone()}}
\item \href{#method-get_attribute}{\code{PopulationSimulator$get_attribute()}}
\item \href{#method-run}{\code{PopulationSimulator$run()}}
\item \href{#method-clone}{\code{PopulationSimulator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method sets the population model, and optionally the simulation function, the sample ID, and any attached attributes listed individually.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulator$new(
  population_model = NULL,
  simulation_function = NULL,
  sample_id = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population_model}}{A PopulationModel object or an inherited class object (can be set later).}

\item{\code{simulation_function}}{Optional name (character string) or direct assignment (assigned or loaded via source path) of the population simulation function, which takes a population model as an input and returns the simulation results.}

\item{\code{sample_id}}{Optional identifier for the simulation sample.}

\item{\code{...}}{Additional parameters passed individually are attached.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new_clone"></a>}}
\subsection{Method \code{new_clone()}}{
Creates a new (re-initialized) object of the current (inherited) object class with optionally passed parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulator$new_clone(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed via the inherited class constructor (defined in initialize and run via new).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
New object of the inherited class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attribute"></a>}}
\subsection{Method \code{get_attribute()}}{
Returns selected named simulator or attached attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulator$get_attribute(param)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param}}{Name of the parameter/attribute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Selected parameter/attribute value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
Runs a population simulator (function), stores the results, and creates a status log entry as a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulator$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representing a simulation log entry with a \emph{successful} boolean and a status message template (with a \emph{\%s} placeholder for the sample identifier).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationSimulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
