% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsSummaryManager.R
\name{ResultsSummaryManager}
\alias{ResultsSummaryManager}
\title{R6 Class Representing a Results Summary Manager.}
\description{
An R6 class to represent a manager for generating summary metrics and/or matrices
from simulation results, as well as optionally from (regenerated) niche carrying
capacities and/or human densities.
}
\section{Super classes}{
\code{\link[paleopop:GenericClass]{paleopop::GenericClass}} -> \code{\link[paleopop:GenericResultsManager]{paleopop::GenericResultsManager}} -> \code{ResultsSummaryManager}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{results_model}}{An object of a class inherited from the GenericResultsModel class for encapsulating and dynamically generating simulation results.}

\item{\code{result_attachment_functions}}{A list of functions for attaching intermediate values to the results model prior to generation.}

\item{\code{summary_metrics}}{An array of names for summary metrics, each of which are calculated as single values for each simulation. These should refer to list names for the summary functions.}

\item{\code{summary_matrices}}{An array of names for summary matrices, each of which are calculated as a single matrix row for each simulation. These should refer to list names for the summary functions.}

\item{\code{summary_functions}}{A list of functions, result attributes, or constants for transforming individual simulation results to single summary metric values stored in the metric data frame, or to matrix rows stored in the summary matrix list.}

\item{\code{summary_metric_data}}{A data frame of generated summary metrics (one row per simulation).}

\item{\code{summary_matrix_list}}{A list of generated matrices of summary results (each having one row per simulation).}

\item{\code{summary_matrix_weighted_averages}}{A list of calculated weighted averages for each of the summary matrices (using the sample data \emph{weight} column).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResultsSummaryManager$new()}}
\item \href{#method-generate}{\code{ResultsSummaryManager$generate()}}
\item \href{#method-calculate_result_attachments}{\code{ResultsSummaryManager$calculate_result_attachments()}}
\item \href{#method-calculate_summaries}{\code{ResultsSummaryManager$calculate_summaries()}}
\item \href{#method-log_generation}{\code{ResultsSummaryManager$log_generation()}}
\item \href{#method-calculate_summary_weighted_averages}{\code{ResultsSummaryManager$calculate_summary_weighted_averages()}}
\item \href{#method-clone}{\code{ResultsSummaryManager$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="new_clone">}\href{../../paleopop/html/GenericClass.html#method-new_clone}{\code{paleopop::GenericClass$new_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericClass" data-id="read_from_rds">}\href{../../paleopop/html/GenericClass.html#method-read_from_rds}{\code{paleopop::GenericClass$read_from_rds()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="get_attribute">}\href{../../paleopop/html/GenericResultsManager.html#method-get_attribute}{\code{paleopop::GenericResultsManager$get_attribute()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="get_message_sample">}\href{../../paleopop/html/GenericResultsManager.html#method-get_message_sample}{\code{paleopop::GenericResultsManager$get_message_sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="get_results_filename">}\href{../../paleopop/html/GenericResultsManager.html#method-get_results_filename}{\code{paleopop::GenericResultsManager$get_results_filename()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="paleopop" data-topic="GenericResultsManager" data-id="save_to_rds">}\href{../../paleopop/html/GenericResultsManager.html#method-save_to_rds}{\code{paleopop::GenericResultsManager$save_to_rds()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialization method optionally copies attributes from a multi-simulation (results) manager, sets any included attributes (sample_data, results_model, niche_k_model, human_density_model, result_attachment_functions, summary_metrics, summary_functions, parallel_cores, results_dir, results_filename_attributes), and attaches other attributes individually listed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultsSummaryManager$new(simulation_manager = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulation_manager}}{Optional simulation manager object of a class inherited from \emph{GenericResultsManager}, from which simulation attributes can be copied.}

\item{\code{...}}{Parameters listed individually.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\subsection{Method \code{generate()}}{
Generates the summary metric data and/or matrix list via the summary functions for each simulation sample, and creates/writes a generation log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultsSummaryManager$generate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Generation log as a list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_result_attachments"></a>}}
\subsection{Method \code{calculate_result_attachments()}}{
Calculates and attaches intermediate values to the sample result model (via the result attachment functions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultsSummaryManager$calculate_result_attachments(results_model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results_model}}{The sample results model, an object of a class inherited from \emph{GenericResultsModel}, to which the intermediate results are attached.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_summaries"></a>}}
\subsection{Method \code{calculate_summaries()}}{
Calculates the summary metrics and/or matrices for the results of a sample simulation (via the summary functions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultsSummaryManager$calculate_summaries(results_model, sample_index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results_model}}{The sample results model, an object of a class inherited from \emph{GenericResultsModel}.}

\item{\code{sample_index}}{Index of sample from data frame.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Generation log entry as a (nested) list, including generated summary metric data and (optionally) matrices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_generation"></a>}}
\subsection{Method \code{log_generation()}}{
Summarizes the log generated within the generate method and writes it to (text and RDS) files in the results directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultsSummaryManager$log_generation(generation_log)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generation_log}}{Nested list of log entries generated via the generate method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate_summary_weighted_averages"></a>}}
\subsection{Method \code{calculate_summary_weighted_averages()}}{
Calculates the weighted averages for each of the summary matrices (providing the sample data has a \emph{weight} column).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultsSummaryManager$calculate_summary_weighted_averages(
  na_replacements = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{na_replacements}}{List of values or functions (form: \code{modified_matrix <- function(matrix)}) for dealing with NA values in each summary matrix (default NULL will ignore NAs).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultsSummaryManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
