% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siberia_raster}
\alias{siberia_raster}
\title{Bison vignette Siberia raster}
\format{
A \emph{raster::RasterStack} object:
\describe{
  \item{dimensions}{21 rows by 180 columns by 1001 layers}
  \item{resolution}{2 by 2 degree grid cells}
  \item{crs}{WGS84 latitude longitude}
  \item{extent}{longitude -180 to 180 degrees; latitude 42 to 84 degrees}
  \item{values}{region defined by 913 cells with value of 1, surrounded by non-region \code{NA} values}
}
}
\source{
TBA
}
\usage{
siberia_raster
}
\description{
A \code{raster} dataset defining the grid cells of the Siberia study region in a temporally dynamic manner for the
bison example vignette.
}
\keyword{datasets}
