% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleopop_simulator.R
\name{paleopop_simulator}
\alias{paleopop_simulator}
\title{Runs a customized population model simulation.}
\usage{
paleopop_simulator(inputs)
}
\arguments{
\item{inputs}{Nested list/object with named elements:
\describe{
  \item{\code{duration}}{Number of simulation time steps.}
  \item{\code{years_per_step}}{Number of years per time step.}
  \item{\code{transition_rate}}{Rate of transition (or fecundity) between generations.}
  \item{\code{standard_deviation}}{Standard deviation applied to transition rates.}
  \item{\code{populations}}{Number of populations.}
  \item{\code{initial_abundances}}{Array of initial abundances for each population.}
  \item{\code{growth_rate_max}}{Maximum growth rate.}
  \item{\code{local_threshold}}{Abundance threshold (that needs to be exceeded) for each population to persist.}
  \item{\code{dispersal_target_k_threshold}}{Target population carrying capacity threshold for density dependent dispersal.}
  \item{\code{carrying_capacities}}{Matrix of carrying capacities (\emph{populations} rows by \emph{duration} columns).}
  \item{\code{dispersal_data}}{Data frame of non-zero dispersal rates and indices for the construction of compact matrices (columns: \emph{target_pop}, \emph{source_pop}, \emph{emigrant_row}, \emph{immigrant_row}, \emph{dispersal_rate}).}
  \item{\code{compact_decomposition}}{List containing a pre-calculated compact transposed (Cholesky) decomposition matrix and a corresponding map of population indices for environmental correlation (list names: \emph{matrix}, \emph{map}).}
  \item{\code{harvest}}{Boolean for utilizing harvesting.}
  \item{\code{harvest_max}}{Proportion harvested per year (note: annual time scale - not generational).}
  \item{\code{harvest_g}}{The \emph{G} parameter in the harvest function.}
  \item{\code{harvest_z}}{The \emph{Z} parameter in the harvest function.}
  \item{\code{harvest_max_n}}{Maximum density per grid cell.}
  \item{\code{human_densities}}{Matrix of human density (fraction) (\emph{populations} rows by \emph{duration} columns).}
  \item{\code{results_selection}}{List of results selection from: "abundance", "ema", "extirpation", "harvested", "occupancy".}
}}
}
\value{
Simulation results as a nested list (as selected):
  \describe{
    \item{\code{abundance}}{Matrix of simulation abundances (\emph{populations} rows by \emph{duration} columns).}
    \item{\code{ema}}{Matrix of expected minimum abundances (\emph{populations} rows by \emph{duration} columns).}
    \item{\code{extirpation}}{Array of extirpation times for each population.}
    \item{\code{harvested}}{Matrix of estimated individuals harvested (\emph{populations} rows by \emph{duration} columns).}
    \item{\code{occupancy}}{Array of number of populations occupied at each time-step.}
  }
}
\description{
Simulates a customized population model, optimized for single-generation transitions
and large populations, across multiple generations and returns simulation results.
Each generational time-step includes:
\enumerate{
  \item Density dependence and harvest calculations
  \item Enviromental stochasticity calculations
  \item Generational transition calculations
  \item Dispersal calculations
  \item Results collection
}
}
